Refactor moved toolbar to own folder, and moved utility functions
