+++ src/toolbar/index.ts
@@ -1,0 +1,2 @@
+export { MdToolbar } from "./toolbar";
+export * from "./types";
